Implements User Usage
User usage works as expected. Given a username, the user controller
searches for all records for that user and display the totals.

Thanks to 21148d5 numbers are properly formatted with correct units and
commas when needed. In addition, routes are being used for friendly
URLs. They take the follow form:
    index.php/user/<user>/

Future work may include additional items such as export to PDF or draw
logarithm graphs of the average file size.

Signed-off-by: Adam Brenner <aebrenne@uci.edu>