Hierarchy Stats / Reports
First version of the hierarchy stats model. The idea with the hierarchy
stats is that we want to take into account all files and folders of a
specific directory tree and count that towards a specific user and
group.

Currently, RBH does not support this. Stats are based off of file
ownership that matches the UID scheme on the system.

An example case of this below:

     Fullpath                    Chown:     User:group
   /fhgfs/aebrenne/file1.txt            aebrenne:staff
   /fhgfs/aebrenne/anotherUser.txt      hmangala:staff
   /fhgfs/aebrenne/fileFromWeb.tar.gz      87291:21323

In the above example, the hierarchy model (configurable) assumes the
second segment in Fullpath (aebrenne) is the user (configurable) and
the group is predefined (configurable). Without the hierarchy model, RBH
would assign anotherUser.txt under the user hmangala and
fileFromWeb.tar.gz to an unknown UID. This will cause disk usage
discrepancy and other reports to be false.

What the hierarchy model does, is ignores the value for both user and
group and assumes all files, folders and symlinks under the directory
tree of /fhgfs/aebrenne/ to be owned by aebrenne regardless of the
acutal ownership reported by the filesystem. It does this for each
directory tree and can be configured via the database.

Upon first public release -- with documentation -- this will be
described in greater detail, including the MySQL table schema.

Signed-off-by: Adam Brenner <aebrenne@uci.edu>