Record and Display Age of Data + Bug Fix + Scheme Changes
When running a cronjob to fetch the stats for both hierarchy and
non-hierarchy system, the timestamp is now recorded into the database.
This timestamp is then displayed to the user offering them context that
results are cached and not live. This completes future work as noted
from commit: 815ea22041

Fixed a small bug when when displaying filesystem usage that would
incorrectly COUNT() the number of files and folders.

Database scheme changes were made for the final push to a stable and
modular release. Scheme now supports both hierarchy and non-hierarchy
within the same scheme. The new scheme allows for switching between the
two be seamless. More on this and documenation to follow soon.

Signed-off-by: Adam Brenner <aebrenne@uci.edu>