Configurations stored in Database
All configurations are now stored in the database. I was toying with the idea
to have configurations done in a file. IMO, I believe storing it to the
database would be easier as certain design choices were made (for example
filesystem database layout) that tie in with configuration option.

Depending on feedback, I can always switch a file based configuration.
Hopefully the documation I write will make it easy for people to use.

Signed-off-by: Adam Brenner <aebrenne@uci.edu>